requirements.txt
asgiref==3.6.0
Django==4.1.7
sqlparse==0.4.3
tzdata==2022.7
psycopg2==2.9.5
## The following requirements were added by pip freeze:
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==39.0.2
defusedxml==0.7.1
django-allauth==0.52.0
django-bootstrap5==22.2
idna==3.4
oauthlib==3.2.2
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.2.0
requests==2.28.2
requests-oauthlib==1.3.1
urllib3==1.26.15
